+++ k/task.h
@@ -2,7 +2,7 @@
 // Created by rebut_p on 28/09/18.
 //
 
-#include "multiboot.h"
+#include "include/multiboot.h"
 
 #ifndef KERNEL_EPITA_USERLAND_H
 #define KERNEL_EPITA_USERLAND_H
@@ -13,6 +13,6 @@
 void launchTask();
 int createTask(const char *cmdline);
 u32 task_switch(u32 previousEsp);
-u32 sbrk(ssize_t inc);
+u32 sbrk(s32 inc);
 
 #endif //KERNEL_EPITA_USERLAND_H
